Entity ID types do not match types in database definition
As Liquibase is used, JPA does not manage database definition.
Within Liquibase changelog, only Anchor and Fragment have BIGINT (Long)
IDs. Other entities have INTEGER IDs.

- Change AnchorEntity/AnchorRepository ID type to Long
- Change YangResourceEntity/YangResourceRepository and
  ModuleReferenceRepository ID to Integer
- Modify datatypes of foreign key references in liquibase changelog to
  use correct types for dataspace_id, schema_set_id, yang_resource_id
- Update test liquibase changelog

Issue-ID: CPS-1676

+++ cps-ri/src/main/java/org/onap/cps/spi/entities/FragmentEntityArranger.java
@@ -58,7 +58,7 @@ public class FragmentEntityArranger {
      * @return a collection of FragmentEntities (trees) with descendants.
      */
     public static Collection<FragmentEntity> toFragmentEntityTreesAcrossAnchors(
-            final Map<Integer, AnchorEntity> anchorEntityPerId, final Collection<FragmentExtract> fragmentExtracts) {
+            final Map<Long, AnchorEntity> anchorEntityPerId, final Collection<FragmentExtract> fragmentExtracts) {
         final Map<Long, FragmentEntity> fragmentEntityPerId = new HashMap<>();
         for (final FragmentExtract fragmentExtract : fragmentExtracts) {
             final AnchorEntity anchorEntity = anchorEntityPerId.get(fragmentExtract.getAnchorId());
