+++ cps-ri/src/main/java/org/onap/cps/spi/repository/YangResourceNativeRepository.java
@@ -1,6 +1,6 @@
 /*-
  * ============LICENSE_START=======================================================
- *  Copyright (C) 2022 Nordix Foundation.
+ *  Copyright (C) 2022-2023 Nordix Foundation.
  * ================================================================================
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,6 +26,6 @@ import org.onap.cps.spi.model.ModuleReference;
 
 public interface YangResourceNativeRepository {
 
-    List<Long> getResourceIdsByModuleReferences(Collection<ModuleReference> moduleReferences);
+    List<Integer> getResourceIdsByModuleReferences(Collection<ModuleReference> moduleReferences);
 
 }
